End to end tests Docs Adding base type Remaining model stuff to JSON Removing added cocoa pods stuff
