+++ GenumKit/Parsers/JSONFileParser.swift
@@ -1,5 +1,5 @@
 //
-//  ModelsFileParser.swift
+//  JSONFileParser.swift
 //  Pods
 //
 //  Created by Peter Livesey on 9/16/16.
@@ -8,7 +8,7 @@
 
 import Foundation
 
-public final class ModelsJSONFileParser {
+public final class JSONFileParser {
   public private(set) var json = [String: AnyObject]()
 
   public init() {}
@@ -17,14 +17,6 @@ public final class ModelsJSONFileParser {
     if let JSONdata = NSData(contentsOfFile: path),
       let json = (try? NSJSONSerialization.JSONObjectWithData(JSONdata, options: [])) as? [String: AnyObject] {
       self.json = json
-    }
-  }
-
-  public func parseDirectory(path: String) throws {
-    if let dirEnum = NSFileManager.defaultManager().enumeratorAtPath(path) {
-      while let subPath = dirEnum.nextObject() as? NSString {
-        try parseFile((path as NSString).stringByAppendingPathComponent(subPath as String))
-      }
     }
   }
 }
