+++ templates/json-model.stencil
